scraper/requirements.txt
requests
kafka-python
schedule
